refactor(angular): update references to utils
